backend/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv

pydantic

ddgs
requests
httpx
beautifulsoup4
newspaper3k
lxml_html_clean

langchain
langchain-openai
langchain-community
langchain-chroma
langgraph

openai
tiktoken

chromadb

tenacity
orjson